+++ components/InfinitePostList.tsx
@@ -1,0 +1,7 @@
+export default function InfinitePostList() {
+  return (
+    <>
+      <h1>InfinitePostList Component</h1>
+    </>
+  );
+}
